vertex_layout::Type is hold by VertexDeclaration
