+++ source/renderer/resource/VertexDeclaration.h
@@ -5,6 +5,26 @@
 
 namespace te
 {
+    namespace vertex_layout
+    {
+        enum Type {
+            Position,
+            Overlay, // Position with Texture Coordinate
+            Model,
+            Partical,
+            PNTB
+        };
+
+        struct VertexDeclarationStream
+        {
+            vertex_layout::Type layout_type;
+            std::vector<GPUResourceHandle*> vertex_buffers; // notice: vertex_buffers might be duplicated here
+                                                            // because multiple attributes may share one vertex buffer
+            GPUResourceHandle*              index_buffer;
+            GPUResourceHandle*              res;
+        };
+    }
+
     class VertexDeclaration : public GPUResource
     {
     public:
@@ -15,6 +35,10 @@ namespace te
         void unload();
 
         void fillStreamItem(ResourceStreamItem& item, Resource* res);
+        void setVertexLayoutType(vertex_layout::Type t) { _vertex_layout_type = t; };
+
+    protected:
+        vertex_layout::Type     _vertex_layout_type;
     };
 }
